action refactor to multi software step 1
